Use a 32-bit index stride for large P3D meshes
The stride selection was inverted, so a mesh with more vertices than a 16-bit
index can address allocated the narrow stride and then narrowed every index
written through the buffer proxy. In a debug build that raised a modal assert,
which looked like the caller hanging rather than failing.

Copilot-Session: 0b20ce7c-3a95-497d-9526-7f85c6a90204

+++ projects/rylogic/view3d-12/src/model/model_generator.cpp
@@ -1609,8 +1609,9 @@ namespace pr::rdr12
 					++vptr;
 				}
 
-				// Copy nuggets
-				m_cache.m_icont.resize(mesh.icount(), big_indices ? sizeof(uint16_t) : sizeof(uint32_t));
+				// Copy nuggets. Meshes with more vertices than a 16-bit index can address need the wider stride,
+				// otherwise writing an index through the buffer proxy narrows it and corrupts the geometry.
+				m_cache.m_icont.resize(mesh.icount(), big_indices ? sizeof(uint32_t) : sizeof(uint16_t));
 				m_cache.m_ncont.reserve(mesh.ncount());
 				auto iptr = m_cache.m_icont.begin<uint32_t>();
 				auto vrange = Range::Zero();
